Use Go style pseudo-versions when there is no semver supplied
The CI workflow uses a semver (e.g. v1.0.0) when you run it using the
workflow_dispatch trigger and supply one explicitly. This is how you
'release' a function.

Today when the workflow runs for a PR or regular main branch build we
default to v0.0.0-<git-short-sha>. This is a very simple way to tag
development builds with no user input required. It's handy that the
resulting package can be correlated to the git commit that produced it
using the git SHA.

A major flaw of this implementation is that newer SHAs do not sort above
older SHAs when treated as semantic versions.

This commit switches us to a simple approximation of the Go
pseudoversions you would see in a go.mod file, as described by
https://go.dev/ref/mod#pseudo-versions. These versions include the git
commit time before the SHA, so newer commits will be considered newer
versions when processing the semver. We also switch to using the first
12 characters of the SHA. This reduces the risk of collision, and makes
us match exactly Go's pseudoversion implementation.

Signed-off-by: Nic Cope <nicc@rk0n.org>